services/rag_api/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
pydantic==2.12.4
langchain-core==1.0.4
langchain-community==0.4.1
langchain-weaviate==0.0.6
langchain-google-genai==3.0.2
weaviate-client==4.17.0
python-dotenv==1.0.1
langchain-huggingface==1.0.1
huggingface-hub==0.36.0
sentence-transformers==5.1.2